cuda version fully reworked

+++ src/common/LPsolution.cpp
@@ -11,7 +11,7 @@ void LPsolution::show()
     std::cout << "phase 2 time:               " << phase2_time << std::endl;
     std::cout << "iterations:                 " << iterations << std::endl;
     std::cout << "x0:                         " << x[0] << std::endl;
-    for (size_t i = 1; i < std::min(int(x.getSize()), 10); i++)
+    for (int i = 1; i < std::min(int(x.getSize()), 10); i++)
         std::cout << "                            " <<  x[i] << std::endl;
     if (x.getSize() > 10)
         std::cout << "                            " <<  "..." << std::endl;
