+++ src/cpu/valuesVector.cpp
@@ -10,7 +10,7 @@ ValuesVector ValuesVector::operator-(const ValuesVector &values_vector) const
     }
     
     ValuesVector new_vec(values_vector.getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
        new_vec[i] = operator[](i) - values_vector[i];
 
     return new_vec;
@@ -20,7 +20,7 @@ ValuesVector ValuesVector::operator-(const ValuesVector &values_vector) const
 ValuesVector ValuesVector::operator-() const
 {
     ValuesVector new_vec(getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
        new_vec[i] =  -operator[](i);
 
     return new_vec;
@@ -35,7 +35,7 @@ double ValuesVector::dot(const ValuesVector &values_vector) const
         exit(1);
     }
     double dot_val = 0;
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
         dot_val += operator[](i) * values_vector[i];
 
     return dot_val;
@@ -60,7 +60,7 @@ double ValuesVector::norm() const
 }
 
 
-size_t ValuesVector::countNonZero() const
+int ValuesVector::countNonZero() const
 {
     double count = 0;
     for (auto a_i : data)
@@ -84,7 +84,7 @@ ValuesVector& ValuesVector::operator-=(const ValuesVector& values_vector)
     }
     
     ValuesVector new_vec(values_vector.getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
         operator[](i) -= values_vector[i];
 
     return *this;   
@@ -100,7 +100,7 @@ ValuesVector& ValuesVector::operator+=(const ValuesVector& values_vector)
     }
     
     ValuesVector new_vec(values_vector.getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
         operator[](i) += values_vector[i];
 
     return *this;   
@@ -116,7 +116,7 @@ ValuesVector ValuesVector::operator+(const ValuesVector& values_vector) const
     }
     
     ValuesVector new_vec(values_vector.getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
        new_vec[i] = operator[](i) + values_vector[i];
 
     return new_vec;
@@ -126,14 +126,14 @@ ValuesVector ValuesVector::operator+(const ValuesVector& values_vector) const
 ValuesVector ValuesVector::operator*(const double& value) const
 {
     ValuesVector new_vec(getSize());
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
        new_vec[i] = value * operator[](i);
 
     return new_vec;
 }
 
 
-ValuesVector::ValuesVector(const size_t size)
+ValuesVector::ValuesVector(const int size)
 {
     data = std::vector<double>(size);
 }
@@ -161,7 +161,7 @@ ValuesVector ValuesVector::operator()(const IndexVector& indexes) const
 {
     ValuesVector new_vector(indexes.size());
     
-    for (size_t i = 0; i < indexes.size(); i++)
+    for (int i = 0; i < indexes.size(); i++)
         new_vector[i] = operator[](indexes[i]);
     
     return new_vector;
@@ -171,7 +171,7 @@ ValuesVector ValuesVector::operator()(const IndexVector& indexes) const
 void ValuesVector::show() const
 {
     std::cout << "Vector(" << getSize() << "):" << std::endl;
-    for (size_t i = 0; i < getSize(); i++)
+    for (int i = 0; i < getSize(); i++)
         std::cout << operator[](i) << " ";
     std::cout << std::endl;
 }
@@ -179,18 +179,18 @@ void ValuesVector::show() const
 
 void ValuesVector::setValues(const ValuesVector& values_vector, const IndexVector& index_vector)
 {
-    for (size_t i = 0; i < index_vector.size(); i++)
+    for (int i = 0; i < index_vector.size(); i++)
     {
-        size_t j = index_vector[i];
+        int j = index_vector[i];
         operator[](j) = values_vector[i];
     } 
 }
 
 
-ValuesVector ValuesVector::operator()(const size_t start, const size_t stop) const
+ValuesVector ValuesVector::operator()(const int start, const int stop) const
 {
     ValuesVector result(stop - start);
-    for (size_t i = start; i < stop; i++)
+    for (int i = start; i < stop; i++)
         result[i] = operator[](i);
     return result;
 }
